Make a move generator file
